+++ example/lib/streams/built_single_stream/load_user_email_from_sqlite_stream.dart
@@ -1,4 +1,5 @@
-import 'package:built_stream/built_stream.dart';
+import 'package:built_stream/stream_annotations.dart';
+import 'package:built_stream/stream_types.dart';
 import 'dart:async';
 import 'package:customized_streams/customized_streams.dart';
 import 'package:example/repositories/user_repository.dart';
@@ -7,7 +8,8 @@ part "load_user_email_from_sqlite_stream.g.dart";
 
 @SingleStream(UserRepository, 'loadUserEmailFromSQLite')
 @StreamResult(String, 'email')
-class LoadUserEmailFromSQLiteStream extends _LoadUserEmailFromSQLiteStreamOrigin {
+class LoadUserEmailFromSQLiteStream
+    extends _LoadUserEmailFromSQLiteStreamOrigin {
   @override
   String get errorMessage => 'Cannot load user\'s email from SQLite';
 }