+++ example/lib/streams/built_single_stream/login_stream.dart
@@ -1,5 +1,8 @@
 import 'package:example/repositories/auth_repository.dart';
-import 'package:built_stream/built_stream.dart';
+
+import 'package:built_stream/stream_annotations.dart';
+import 'package:built_stream/stream_types.dart';
+
 import 'dart:async';
 import 'package:customized_streams/customized_streams.dart';