+++ example/lib/streams/built_single_stream/notify_user_session_to_server_stream.dart
@@ -1,4 +1,5 @@
-import 'package:built_stream/built_stream.dart';
+import 'package:built_stream/stream_annotations.dart';
+import 'package:built_stream/stream_types.dart';
 import 'dart:async';
 import 'package:customized_streams/customized_streams.dart';
 import 'package:example/repositories/user_repository.dart';